chore(alchemy): export types
