+++ apps/indexer/package.json
@@ -16,6 +16,7 @@
     "@dfuse/client": "^0.3.21",
     "@repo/supabase": "workspace:*",
     "@repo/trigger": "workspace:*",
+    "@repo/alchemy": "workspace:*",
     "@sentry/integrations": "^7.114.0",
     "@sentry/node": "^8.19.0",
     "@sentry/profiling-node": "^8.26.0",
